Add "Negative" to temp check
The temperature check currently has:
- two positive reactions
- following, which means "I am curios to see how the committee feel, but I don't have a stance"
- confused, which means "I do not understand what this is about"
- unconvinced, which means "I do understand it, but I am not convinced by your motivation"

This PR adds an explicitly negative option. So far, delegates had to pick one between unconvinced and confused to mean negative, but adding an explicit negative option makes it easier to distinguish between "I don't care", "Useless, but not harmful" and "I don't think we should add this to the language".